The first look at my GetMileage application
I created this application to handle a couple of repetitive tasks that
support my habit of keeping details of all my daily workouts in plain
text log-files. It was also a good chance to practice File I/O and
building a GUI with Swing. The application adds up and displays the
miles I've run over a specified period of time, and creates a new
log-file for a given year.