Incorporated java.time for LogBuilder class
I rewrote LogBuilder using the java.time functionality. This makes the
code more readable, and simplifies the user interface because now the
program automatically determines leap years and the day of the week to
start the year. Also found and fixed a bug that led to changes to the
"working directory" failing to carry over to writing new log files.